made model understand roles
