+++ src/types/index.ts
@@ -1,0 +1,18 @@
+export type Role = "Patient" | "Provider";
+
+export interface Message {
+  sender: Role;
+  inputLang: string;
+  targetLang: string;
+  original: string;
+  corrected: string;
+  suggestions: string;
+  translation: string;
+}
+
+export interface SpeechToTextProps {
+  onTranscript: (text: string) => void | Promise<void>;
+  isListening: boolean;
+  setIsListening: React.Dispatch<React.SetStateAction<boolean>>;
+  inputLang: string;
+}
